MULE-18516: TCP Connector is only reading the first packet of a multipacket message, truncating the payload when using xml-protocol
Removed the check for the full usage of the buffer as a requirement to contnue reading. This is not guaranteed to be always the case by InputStream interface so if the buffer was not used in its entirety it does not mean there is no more data to read.

Changed a test that claimed that when the input stream is slow then the data should not be read entirely.